requirements.txt
streamlit==1.27.0
openai==1.5.0
pydub==0.25.1
numpy==1.24.3
sounddevice==0.4.6
soundfile==0.12.1
python-dotenv==1.0.0
watchdog==3.0.0
websocket-client==1.6.0 
httpx==0.27.2
moviepy==1.0.3
pytest